requirements.txt
mysql-connector-python~=8.0.27
opencv-python~=4.5.4.58
Flask~=2.0.2
Werkzeug~=2.0.2
Pillow~=8.4.0
pytesseract~=0.3.8
face_recognition